requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
discord-py-interactions==5.11.0
discord-py-slash-command==4.2.1
discord-typings==0.7.0
discord.py==2.3.2
emoji==2.10.1
frozenlist==1.4.1
greenlet==3.0.3
idna==3.6
multidict==6.0.5
numpy==1.26.4
pandas==2.1.0
lb_tech_handler==1.0.0
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
six==1.16.0
SQLAlchemy==1.4.17
tomli==2.0.1
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
yarl==1.9.4
